Fix state freezing
Sometimes the state would freeze during regurlar usage of the program.
Though we were unable to track down the exact source of the problem, the
simple truth is that the state mutexing does not follow the conventions,
as it was written before the conventions.

The only way to be certain about a fix is to completely overhaul how the
state manages mutexes, so in the meantime I've just made some dangerous
changes that will allow race conditions thorugh sometimes. They will
also probably prevent deadlock, though since we can't consistently
trigger the error I'm not certain.

This unfortunately goes on the list of things that we hope to fix during
architecture week.