remove miner interface from modules
Reason: it's unlikely that another module is going to need to do
anything with a miner. We don't want modules to be able to turn the
miner on and off or set the threads, which pretty much just leaves
solving a block, but even that seems unreasonable. So the interface was
deleted and deemed uneeded.