TempDir removes old dir before returning
Cleaning up old test data is a problem. We can't simply rm -rf
/tmp/SiaTesting in the Makefile, because that's OS-specific. So
TempDir will now call os.Remove on the path that it returns. The
only downside to this approach is that it will only clean up the
paths used for a given test, so if you don't test everything you
may wind up with old test data. But is that really a downside?