add a constant for notify buffers
When testing, it is important to be certain about the state of
synchronization of all of the modules. Often, this means waiting for
some updates to propagate. Some module calls (in particular, the api
call to spend coins) result in multiple updates being released at the
same time. The buffer is to make sure that all of the updates can be
seen and the test can block until the updates have propagated to all
modules.

Right now the buffer is set to 3, which is sufficiently large for all
operations. In the future, operations may be added that require a larger
notify buffer.