switch back to using unsafe in miner
The miner was switched back from incrementing bytes individually to
using unsafe to increment an int represetation of the nonce. This was
mostly for clarity.

Additionally, 'SolveBlock' was switched to using the more efficient
mining implementation, resulting in probably a 3x speedup for finding
blocks. During testing, finding blocks was something of a bottleneck,
this might help with that.