Replaced 'spend' with 'send' for siafund too
An issue came to mind about this when grepping:

The word 'Spend' is so similar to the word 'send' yet different in
connotation. 'Spend' is also very inextricably injected all across
Sia in variable names, functions names, and documentation
(including the whitepaper) that we should keep an eye out for how
to keep our terminology consistent here... and everywhere else.

Also made a couple corresponding comment corrections