repair impossible storage proof bug - soft fork
Due to inconsistency with how the crypto package treated padding,
certain storage proof requests were actually impossible to complete.
When creating the merkle root, the last segment of a file has not been
padded out, but instead treated as a smaller array. Whether this
decision was good or bad, there are a bunch of file contracts on the
blockchain right now that don't have padding on their last segment.

The storage proof struct, on the other hand, pads out the final segment
so that it is 64 bytes in size. In the majority of cases, this does not
cause any issues. However, if the random segment that gets selected is
the final segment and the final segment is less than 64 bytes in size,
then the storage proof is actually impossible to create because the
consensus code adds padding at the end when trying to verify the code.

This commit adds a hardfork that makes storage proofs on the last
segment legal again.