implement walletSeedHandlerGET
This actually resulted in a modification of the entropy-mnemonics
library, and a modification to the wallet interface. Previously, the
master key was required to load all seeds. That requirement has been
removed, and I believe in doing so more safetly has been added to the
way that seeds are determined.