update constants
the devnet default block target was reduced by a factor of 5 to account
for the lack of multithreading and the newly common use of the race
detector during developer testing.

The default target for testing was raised from 4 hashes to 2 hashes,
because there are a lot of tests now,  tests get run frequently and with
different genesis blocks each time, and the hash doesn't match code gets
plenty of coverage. Less hashing means faster tests, especially when
building out many many blocks.