remove deleteNode from consensus
This does mean that some cruft can build up in the consensus package,
but it's cruft that is very expensive for an attacker to generate. It
does not increase the vulnerability of the node to computational
attacks.

Removing deleteNode decreases the amount of code and the code complexity
of the consensus set, which makes it less likely that there is a bug.