add txn purging to the renter
If negotiation failed at some point, the transaction was left in limbo -
outputs had been consumed according to the wallet, but actually had not
been consumed since the transaction had failed. The transaction builder
now includes a purge function that allows for outputs to be 'dropped'
from a transaction builder in the event of a failed transaction.

And the renter will now purge any in-progress transaction if negotiation
fails, so that any outputs can be restored.