more agressive height checking for types package
I realized that the height in the types package would not be set
correctly during large reorgs of the consensus set - only during the
first reorg. I don't think it's likely to be a problem on the live
network because reorgs are fairly rare, but the code is more correct
now.