hopefully fix muxado panics
I think passing a nil net.Conn was the culprit. It was deceptive
because the test wasn't always failing during the test containing
the nil. I suspect muxado was cleaning up the connection after the
test had already returned, so the panic was triggered in the middle
of the next test.