add on-disk changelog
The on-disk changelog keeps a persistent changelog for the consensus
set. Subscribers will be able to use this changelog to avoid needing to
be sent the entire blockchain + diff set every time siad starts.

The changelog is now updated at the same time that the changes to the
consensus set get committed.

Other organization was also applied.