Add preprocessing for the configuration to Siad
Specifically, preprocessing was added for portnumbers in siad.
Previously, a port number provided as an integer resulted as an error, a
prefixed ":" was required. Now, it is okay to provide just an integer as
the preprocesser will clean up the number.

This was done by adding a function `preprocessConfig` which can be
extended to clean up other configuration values in the future.

fixes #650

---

'siad' previously had no testing. In this commit, a test package was
added for 'siad' so that `preprocessConfig` could be unit tested.