generate individual keys and save them
Previously, the hostdb deterministically generated signing keys
by combining the host's key with 32 bytes of "master" entropy.
However, the host's key was not being saved to disk anyway, so
this key could not be recovered for future revisions. The hostdb
now generates a new randomly-seeded key for each contract, and
saves it as part of the hostContract struct.