more pure startDaemon
The config global variable is only touched by the cobra package now, and
'startDaemon' takes in a Config object as input instead of looking at
the global object.

The tests for 'processConfig' were borken apart, with one set of tests
dedicated to handling 'processNetAddr' and the other set of tests
checking that APIaddr, RPCaddr, and HostAddr are all properly being
handled.

GOMAXPROCS is no longer set manually, because the golang defaults have
been improved. http://tip.golang.org/doc/go1.5 "By default, Go programs
run with GOMAXPROCS set to the number of cores available; in prior
releases it defaulted to 1."