move unlock hash to last check in explorer
the unlock hash needs to be checked last because otherwise it is
possible to maliciously create collisions that would hide prevent people
from being able to find other types of hashes in the explorer. This does
mean that they will not be able to find the collided unlock hash through
the explorer directly, but that shouldn't matter because it the error
can only be caused maliciously anyway.

Long term, there may be an improvement to more thoroughly handle this
edge case.