add support for historic block facts to the explorer
the previously tracked a bunch of statistics about the current state of
the consensus set. Now, those statistics are saved for every block, that
way historic data can be queried directly from siad.

Some of the 'active' facts are not always going to be correct, due to a
limitation in the subscriber model that prevents diffs from being sent
to the explorer one block at a time. A fix would require adding a method
to the consensus set that returns the diffs for a specific block, and
then applying diffs to the active set after each applied block and
reverted block as opposed to as a group at the end of the update.