use host pool abstraction in renter
Host Pools allow the renter to easily select unique hosts for a
given chunk, and close the connections to all hosts without needing
to maintain a separate list of them somewhere. Allowing the HostDB
to manage the Host Pool means that it can efficiently reuse existing
connections and/or file contracts if available, and form new
contracts as needed.