Expand the explore api
A few fields got added to the explorer api, and a few got changed that
will break compatibility with the current website. A lot of redundant
code got elimiated in the process.

The fields don't all fill out correctly yet, because the explore module
itself is still missing features. That is coming next.