expand the explorer api fields
The explorer now includes fields to connect siacoin inputs and siafund
inputs with their corresponding outputs, that way the explorer can
display the unlock hash and value associated with the input when
digesting a transaction.

A similar thing was done for file contracts. Now, the file contract, the
most recent blockchain revision, and the storage proof can all be
queried from the explorer, and when digesting a storage proof in a
transaction the corresponding outputs can be displayed. Note that the
explorer is currently blind to all missed file contracts.