add rpcRenew
This is a new Host RPC that allows a renter to renew an existing
contract without having to reupload the contract data. The RPC
protocol is identical to standard contract negotiation, except
that the old Merkle root is used instead of a blank one.
As a result, the negotiation logic has been factored out into a
separate function, used by both rpcUpload and rpcRenew. It is yet
unknown if there are subtleties to rpcRenew that make this reuse
a bad idea. It makes implementation easy though!