add UniqueHosts test
This test is pretty lacking because the pool needs a real HostDB
to function properly. The best route is probably to have pool take
an interface instead of a HostDB; all it needs to provide are two
functions: `randomHosts` and `newHostUploader`.
Well, I guess it also needs locks...this could get ugly. Not going
to spend too much time on it for now.