Clean up FileContractHistory for modules.Explorer
Previously, the file contract, all of the revisions, and the storage
proof were returned, along with 3 bools indicating the existence of each
element. It was decided that the storage proofs do not contribute much
of value, because a storage proof just consists of a parent (already
known, as it is in the input to the function), and a bunch of
cryptographic hashes that are completely meaningless from a human
perspective.

Now, the storage proof is not returned, and because the file contract
revisions are returned as a slice that can have the length checked, a
bool is not returned indicating their existence.