eliminate intentional underflow in miner
The miner previously performed an intentional underflow so that the
genesis block would set the height at zero. By adding some code to the
update function, the need for this underflow was eliminated.

While it makes the update function slightly uglier, and slows down the
program as a whole by a trivial amount, this overhead was deemed the
lesser evil compared to performing intentional underflows in ways that
are not necessarily intuitive.