set testing genesis timestamp to earlier value
While performing some of the timestamp testing, there was a hidden
constraint that the tests could only pass if they were run at least 7
seconds after the initialization of the binary.

To remove this dependency, the timestamp was set further in the past.

To allow flexibilty for furhter testing, the timestamp was set 1e6
seconds in the past. This keeps things nondeterministic (desirable for
hitting a wider potential set of problems) while allowing people to
experiment with the timestamp code.