copy renewed data to new file
David pointed out that two contracts can't point to the same file
on disk, because a revision to one contract would affect the other.
So we are forced to duplicate the data, despite the inefficiency.
Perhaps someday we could optimize this by using a symlink initially,
and performing a full copy only if one of the contracts is revised.