fix contractSize bug
contractSize did not take into account encryption overhead, causing
contracts to be too small.
There is a bigger question of whether f.pieceSize should report the
size pre- or post-encryption. I think there's an argument to be made
for the latter, since the download logic has to tiptoe around it as
well.