check AllHosts in offlineChunks
checking only ActiveHosts is bad because the host in question may
not have been scanned yet. This algorithm should do a decent job,
but testing is required to confirm so.

I'm happy that this didn't require adding a new method to the hostdb,
but that may change if a more sophisticated algorithm is warranted.
For that reason, I changed the argument to offlineChunks to a hostDB.
This should make it easy to swap out the current algorithm for a
method like hdb.IsOffline().