gross refactor
A lot of code needed to be changed, and unfortunately I made a huge mess
out of the refactor. So a bunch of things changed and I don't have a
clear memory of where I started from.

One of the sigificant things that has been added is that the host will
now follow the consensus package and resubmit transactions if they fail,
and remove obligations if someone tries to double-spend after making a
file contract.

The rpc counters have been upgraded so that all operations on those
values are atomic.

As of this commit, the testing is probably largely incomplete, and code
coverage may be deceptive. While the code is hit, the test may not
actually correctly verify that the code ran correctly. Future commits
will improve the testing.

obligations.go has been created, which defines a bunch of operations
related to file contract obligations. Obligations now need to store the
transactions associated with the obligation.

Compatibility has been broken. File contracts loaded from 0.4.8 will be
handled incorrectly by the new logic, so some form of compatibility will
need to be added.