add sanity checks to the logger
The code was originally written to add thread safety to the logger, so
that calls to 'Write' and calls to 'Close' could be made concurrently,
and so that 'Close' wouldn't be called before the most recent 'Write'
had time to finish.

Then it was determined that 'Close' shouldn't be called unless all calls
to 'Write' have finished, and that the standard library logger was safe
enough to begin with. While the new code is technically not necessary,
it does add some sanity checks to make sure that the logger is being
used and closed properly.