threading and closing sanity in host
The host should now do safer threading and close cleanly and correctly.
Because many threaded functions use the logger, and 'Close' closes the
logger, it's necessary to make sure that all active threads have time to
terminate before the close completes. To accomplish this, a
'resourceLock' was added, which is readlocked by any thread that needs
access to host resources such as the logger. The host will writelock the
resource lock before closing, guaranteeing that it is not being held by
any other threads. Then the host is marked as closed so that any threads
waiting on the lock will recognize that the required resources are no
longer available.