better action item handling in host
Previously, the host would queue up a new action item every time there
was a revision, which resulted in things getting queued all the way down
the line for the host. This resulted in many storage proofs being
created.

One of the core problems was that the queue was an array instead of a
map, which means that the same action could be queued multiple times at
the same height. This was fixed by creating a double nested map which
only allows a contract obligation to be at each height once.

As a result, fewer storage proofs need to be created.