ask for correct number of hosts
previously we asked for the total number of pieces minus the current
number of hosts, but this did not produce the correct number when
trying to replace offline hosts; f.chunkHosts would return one host
for each piece, so the result would be 0. The sensible approach is
to simply ask for one new host for each bad piece.