add blacklist to HostPool
Also always ask for at least 10 random hosts, otherwise we risk
picking the same 1-3 hosts every time and failing to form contracts
with them.
There is little harm to asking the hostdb for more hosts; the only
reason we don't ask for the entire set is to avoid getting the
"dregs," i.e. hosts that are poorly ranked.