add a toggle for new contracts to the host
When accepting new contracts is enabled, the host will accept new
contracts. When disabled, the host will not accept new contracts. The
host will not accept revisions or renewals either. The host will accept
download requests, settings requests, and uptime requests.

An 'accepting contracts' flag has been added to the host settings
object. This breaks compatibility with v0.5.0-rc and v0.5.0-rc2 hosts,
but not with anyone else.

This may confuse the renter, because the renter tells that a host is
online by checking the 'Settings' RPC, which the host will correctly
respond to.  The renter has a blacklist, which should help with some of
the issues that might arise, but the renter should really add a way to
check for the 'accepting contracts' flag in the host settings object.