encode multiple shared files properly
Previously, each file would create, write to, and close its own
gzip stream. This causes problems when multiple files are written
to the same stream, because the gzip standard practice is to reuse
the same gzip writer for multiple files. This means the resulting
.sia will have only one footer (checksum). Size may also be reduced
because the same dictionary is used for multiple files.

This should not introduce any compatibility concerns, because prior
to v0.5.0 the API did not support sharing multiple files.