improve pricechecking in host
The price checking in the host made some assumptions that weren't
correctly asserted (that file contracts would have 0 bytes). When those
assumptions changed, some of the considerContract logic broke. Namely,
it became possible to swindle the host out of money by submitting new
file contracts that didn't pay for themselves, and by renewing without
paying for the renewal.

The price checking has been made more strict, but may still be broken on
the renter side since the two are symmetric.

Some similar changes were made in conseiderRevision, the price checking
was incorrect. When the price of the host is changed, the revision
should only need to pay the new price for the new diff. The host however
was expecting the whole file contract to be refinanced. This is now
fixed for the host, but should be inspected in the renter as well.

Testing for all of these properties is in progress.