prepare the constants for a full fledged release
For v0.5.0, The renew period happens 6 weeks out. Files are uploaded for
60 days at a time, which means the renew overhead is about 30%. This is
considered acceptable for consumer use. Enterprise use can significantly
reduce the renew window, perhaps as low as 1 week or even 3 days.

The host will reject all file contracts lasting more than 180 days, or 6
months. This is in preparation for a future release, 1.0, which will
have support for uploading files 6 months at a time.

Redundancy is set to 4 data pieces and 20 parity pieces. This is a total
redundancy of 6x, which is very high. We want to prove reliability. This
is a bit of a risk, using 24 hosts when the number of total hosts means
there is potential for price abuse. We picked these numbers expecting
the total volume of hosts to increase with the release, as we are going
to be doing a significant amount of marketing.