add to allHosts in insertHost
delaying the insert to allHosts until after the scan completed
prevents efficient deduplication. That is, during startup, insertHost
may be called many times on the same address before the address is
actually scanned. This results in an undesirable proliferation of
goroutines. For reference, at the time of this commit, there are
126,000 host announcements in the blockchain, but only 684 unique
addresses.