Add Redundancy to FileInfo
Redundancy is calculated by dividing the total amount of data uploaded
by the padded size of the file. This results in the average chunk
redundancy across all chunks. This does not necessarily correspond to
the availability. Specifically, a redundancy >= 1 does not indicate the
file is available as there could be a chunk with 0 redundancy.