add Uint64() method to types.Currency
A future commit makes use of this function call. An error is returned if
there is an overflow. Though a panic could have been called, it is
believed that this method will be used heavily on currencies that could
potentially be manipulated by untrusted input. Rather than require that
all untrusted input be handled to check for currency-overflow causing
problems, and error is returned by the currency type.