Check error returned from CS's PersistentSubscribe
Because all modules (but the miner) subscribe to the consensus set with
an initial change id of the genesis block, it's only possible for an
error to be returned by ConsensusSetPersistentSubscribe right now if
computeConsensusChange fails. In the future these modules will use a
more recent ConsensusChangeID which could return make
ConsensusSetPersistentSubscribe return ErrInvalidConsensusChangeID if
the module and the consensus package become desynchronized.