Add Send1Blk RPC
The RPC enables a peer to request a specific block id from another peer.
The requested block is received and submitted for acceptance into the
consensus set. It is used in conjuction with the threadedReceiveBlock
method which returns an RPC func that requests the specified block id.
This RPC is used to enable header broadcasting instead of block
broadcasting.