move hostPool to renter
This may seem unnecessary now, but eventually the hostPool will
want access to the Renter so that it knows which sectors are
available in a given contract. (The contractor knows which sectors
are unused, but not which sectors contain expired data.)

Support for forming new contracts in uniqueHosts was also removed.
This makes it completely useless until the contractor starts forming
contracts independently of the Renter.