finally, some sanity
While most of the build is still broken, all of the host tests that
remain are in working order. The sector stuff and the storage folder
stuff have been greatly enhanced, and while there's still some things
missing, the comprehensive smoke test is now working.

A much larger description can be found in the still-chaotic comments.
There's a decent amount of work before this becomes a readable PR, and
even more work before it's a production-ready PR. But we're out of the
darkenss at this point for the storage folders feature.