Create StorageFolderManager interface
The storage folder managment functions of the host have been broken into
a separate interface. This helps to modularize things and separate the
concerns of the host. The internal implementation of the host is still a
merged implementation, but it should not be very difficult to break the
manager into something that's completely separate.