add smoke testing for disk statistics
Some lightweight checks were added to the smoke testing to make sure
that sucesssful disk reads and disk writes are being recorded by the
storage folders. There was not a ton of rigor placed on getting the
exact numbers correct, because the numbers in production are likely to
be both large and inconsequential. It's more important that the numbers
are going up when they should be going up, and aren't going up when they
shouldn't be. And even more importantly, the failure statistics are
going up when there are failures and are not going up otherwise.