better bandwidth control for host interface
The host interface has been given improved bandwidth control. The idea
is that there are a lot of potential external factors that could
influence what the bandwidth limits of the host should be, and rather
than try to create an interface which could cover all of them, we'd
create something that's flexible enough for an external master process
to adjust the limits to respond to potential external conditions.

Maybe in the future the interface will be expanded, but more likely an
extra module will be added which can do all of that management.