fix miner height calculation
In addition to fixing the height calculation, a bunch of logging and
debugging statements were added which will more forcefully catch errors
that the miner is experiencing, printing them to stderr and logging
them.

Additionally, more testing was added to the miner to make sure that all
of the obvious situations where there might be a height failure in the
miner have been accounted for.