modify crypto/signatures.go
Got rid of two functions which were not in use in the project, and
improved testing coverage for signatures.go. Though it's not 100%, the
final error condition is entirely impossible to hit, and is only checked
as a buffer against potential extensions to the library in the future to
support more signature types.