additional debugging and sanity for miner
Though I thought I had fixed the miner height bug, a user reported an
error message that suggested the problem had not been fixed. It also
appears like the problem may be more deep-seated than originally
suggested, instead of being a problem with the miner module it might
actually be a problem with the ConsensusChange propagation, which means
that more modules than just the miner would be affected.

I've added more sanity checks and logging messages to the miner, we
should be closer to getting to the bottom of this.