install and release now follow pkgs variable
Previous to this commit, 'install' and 'release' would build all of the
package regardless of the pkgs flag. While in most cases they will be
installing all of the packages, I did run into a few issues where I only
wanted to build a subset of packages.

This inconsistency has now been resolved.