fix up host announcement protocol
Also move all of the negotiation specific structs, variables, and
functions into their own file, because the host file had become
cluttered and unreadable.

The host announcement protocol is now protected by a public key and a
signature. Hosts, instead of being recognized by their IP address, will
now be recognized by their public keys. This means that hosts have
methods for migrating ip addresses while still being recognized as the
same host.

Some of the code for creating and parsing announcements got moved into
the modules package, because when it is there it can be tested
side-by-side, without an integration test. It also means that the
implementations of the encoding and decoding are right next to
eachother, and that all modules that need to either encode or decode the
host announcements can use the same code for doing so.