a mess of host improvements
As the host is nearing completion, I've become a bit more perfectionist
with the coding. Consequently, while working on one thing I went off and
fixed a few others. The result is that this commit has a whole bunch of
changes that really should have been spread out better.

Approximate list:

+ Added the FinancialMetrics RPC, and also improved the financial
metrics
+ Changed the way that the host manages the netAddress. Instead of
having the net address in the settings and the internal net address be
the same, they are now different. The internal net address is the
address that is discovered when doing a hostname lookup. It is not used
if the settings.NetAddress address is set.
+ Some extra cleanup and close attention was payed to file contract
forming.