recent revision now supports challenges
The host will now propose a challenge to the renter, and the renter must
prove that they own the public key that they will need to perform the
file contract  revision. This is to prevent a DoS where an attacker
could otherwise lock a file contract revision without being able to edit
it.