create storagemanager
StorageManager is an interface which can do all of the storage
management functions. People have had many ideas for how to improve
storage management, and now there is a way for people to create their
own storage managers, or use modified storage managers.

Breaking the coupling between the code leaves much more room for people
to experiment without needing to worry about interfering with the rest
of the functions of the host.