attempt to fix a nondeterministic build failure bug
Sometimes, especially on slower systems like Travis, the consensus
tests, and in particular the synchronization tests, will fail. This
commit attempts to resolve one of the nondeterministic failures relating
to a block not propagating fast enough. Now, the test will wait a much
longer period of time, and do iterative waiting, before declaring that a
network communication has failed.