fix locking in consensus package (#1093)
There were a few locking bugs in the consensus package, places where the
database was being accesed without the readlock.

There was one critical place in particular, which was using a readlock
followed by a writelock, when it really needed to be using a single
writelock -> demotelock. This place is likely the cause of the consensus
desynchronizations between the miner + consensus package, has likely
impacted other modules, and since fixing it I have not been able to
trigger the behavior in question.