repair broken test
A test was experiencing nondeterministic build failures. Mostly during
short tests. The problem appears to be that sometimes a block would be
mined and added to one of the consensus sets before the other had an
opportunity to go through the startup routines, so when it did finish
startup it would synchronize AFTER the block had been mined, screwing up
the test, which relied on the fact that the first consensus set had not
received the block.

The test has had a sleep added to it, and it has moved to the long
tests. Failures from this test should become much more rare, if they
don't stop occurring altogether.

Fixes #1063