makefile improvements
'make dependencies' would fail after upgrading go if you did not
reinstall the standard library. This hsa been fixed by adding
'go install ./...' to the head of the 'make dependencies' command, which
slows things down by a tiny fraction but prevents a class of errors
which may frustrate users.

The '-a' flag has been removed from the makefile, because as of a recent
version of go it is no longer necessary to have that flag in order to
have up-to-date standard library packages. Further, having the '-a' flag
would cause permission issues on systems where $GOROOT or $GOPATH was
owned by root. Removing the flag removes the problem.