simple hostdb fix
The persistence on the hostdb was doing weird things. Turns out, this
was the fault of a `for _, object := range objects` loop, where the
address of 'object' was being used to store the value in the map. Each
iteration of the loop though, 'object' has the same address, which means
the pointers all ended up pointing to the same place, meaning all of the
original data was lost and you ended up with a bunch of the same host
instead of a bunch of different hosts.

It was easy to fix by switching to a `for i := range objects` loop and
then grabbing the pointer of 'objects[i]' instead of grabbing the
pointer of 'object'.