Tpool now recognizes partially confirmed tsets
The transaction pool previously would reject transaction sets if some of
the parent transactions had been confirmed on the blockchain, preventing
the children from ever being accepted or making it onto the blockchain.

Now, the transaction pool will prune any parents which have been
confirmed on the blockchain. Persistence was added to the transaction
pool to make this possible.