clean up TestIntegrationRelaySynchronize
An issue has reported that there was a nondeterminisitc failure related
to a block not extending the longest chain. The line of code responsible
has changed since the issue was created, obfuscating the problem.

I was unable to find an issue, though I did do some cleanup. Hopefully
there will be more clear errors next time there is a problem.