special case for unacceptable renew heights
The host will reject renewals that do not extend the WindowEnd.
This is a problem if the user calls SetAllowance with a smaller
period. For now, we will simply set the allowance and return
without forming any contracts. The previously-formed contracts
will be renewed when they are processed by managedRenewContracts